requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
langchain
openai
chromadb
sentence-transformers
python-dotenv
